Bicycle drummer 0.1
Bicycle drummer : uses an arduino nano, a dreamblaster s1 synth module,
and a battery powered speaker to create a Bicycle drummer. It plays a
rock drum pattern, synchronized to the tempo of the bicycle wheel.  The
Arduino measures the period of the reed switch pulse, and adjusts the
tempo of a drum sequence accordingly.  If the period grows longer than 5
seconds, the sequence is stopped.